requirements.txt
opencv-python==4.1.2.30
lxml
tqdm
numpy<1.19.0
tensorflow
absl-py
easydict
matplotlib
pillow
